Add initial checkpoint and restore functionality. (#178)
This allows a user to run a simulation for a number of time steps and
save the state of the simulation. The state can then be restored to
continue the simulation.

This is,for example, useful to compare the impact of an intervention
based on different parameters. The simulation can be run up to the point
of an intervention, when the parameters' value don't matter. The
simulation can then be resumed multiple times but with different
intervention parameters each time.

For now, only variables are supported. Events will be added next.