Fix broken pkgdown index.
I hadn't added the new types and functions to the pkgdown index, which
broke the site rebuild. This didn't get caught by CI because pkgdown was
only running on the master branch.

Added the missing index entries and changed the CI to run pkgdown
everywhere. It only does the actual push when on the master branch.