Check if the schedule is empty before reading it. (#184)
The `Event::should_trigger` method uses `.begin()` on the set of times
at which it needs to fire to compare that value with the current tick.
However it was missing a check in the case of an empty set. When that
happens, `begin()` returns a pointer to uninitialized memory and we we
reading garbage out of it. If that garbage happens to be equal to the
current time step, the event would fire spuriously, even though it
wasn't meant to.

The fix is straightforward: we just need to check if the set is empty
before reading from the set. If it is empty then the event never
triggers.