Switch Bitset from an R6 class to a named list.
R6 object have significant overhead, especially when instantiating the
objects. While this is acceptable for long-lived objects such as events
or variables, bitsets are created and destroyed very regularly during
simulations.

We can replace our use of an R6 class for `Bitset` with named lists that
are intended to look and feel just like the original API, but which
significant performance improvement. The reference semantics provided by
R6 don't matter in our case, since all mutability happens behind the
external pointer.

On malariasimulation, I get a 30-35% performance improvement when using
this new implementation of Bitset on population sizes under 10k, and
about 10% speedup at 100k.

The object-oriented named list based interface still adds a bit of
overhead compared to using the externalptr and Rcpp functions directly,
but doing so requires intrusive changes in the use site.