Support vectors of logicals as an argument to filter_bitset. (#199)
In addition to a vector of numerical indices and another bitset, the
filter_bitset now accepts a vector of logicals (ie. booleans). Each set
bit of the bitset is included in the return value only if the
corresponding value in the logical vector is TRUE.

Given `v` a logical vector, `filter_bitset(b, v)` is equivalent to the
already supported `filter_bitset(v, which(v))` but is more performant.
The new semantics also align with subsetting of R vectors, which
similarly supports either a vector of indices or vector of logicals.